requirements.txt
Django
requests
django-tinymce